BACKEND/requirements.txt
fastapi==0.115.6
httpx>=0.28.0
uvicorn[standard]==0.32.1
python-multipart==0.0.20
websockets==14.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.10.0
google-cloud-aiplatform>=1.74.0
google-genai>=1.0.0
groq==0.13.1
sentence-transformers==3.3.1
numpy==2.2.1
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
aiofiles==24.1.0
PyPDF2==3.0.1
kokoro>=0.9.4
soundfile>=0.13.0
